Improve Result
To make Result more usable, several methods has been added:
- assert() to explicitly throw if something went wrong
- error() to return an Error if something went wrong or null otherwise
- merge() to merge another Result, propagating errors if any or replacing values
- mergeErrors() and mergeError() to create an errored Result that merge errors of current Result if any
- extractWith() to return from a given callable called passing wrapped value of the Result if not errored, or return a default otherwise

Because Error and ErrorCollector are now used in the result it make sense that those are moved to the outer scope where Result is, instead of being inside the /Query namespace. Even because they have nothing strictly related to query.