Improve transaction: callbacks will now receive Result
The workflow with "signal" callback and ErrorCollector was not really usable, considering that inside transactiona consumers will likely deal with Result instances coming from queries.

For the same reason a Query\ResultSet::toResult() has been implemented, so it will be easier to deal with such objects inside transaction callbacks.

To avoid confusion, some methods on Select and ResultSet that contains "result" are renamed to make clear have nothing to do with Result class.