Fix internal cache keys are not clean up
To prevent expensive cache keys calculation an internal
cache is used.

When the cache for one or more tables is deleted the internal cache keys
does not get removed and this cause the same key to always be returned
once it's calculated.

This change automatically clean up the cache keys depending on the table
given.

If one table is given all of the keys where the substring of the table
is found will be deleted, if more than one table is given then we remove
only the exact match for the keys.