Be more permissive on clean cache tables
Do not be strict on the cache key based on the tables list given.

Previous solution removed all occurrences of a table name from the
`tableKeys` only if one table name was given and in case of more than
one table name we removed only the exact match.

This has various drawbacks, bad communication and strictness rules
are not a good thing because we cannot know all of the possible
key combinations in advance.

For the aforementioned reasons I made some changes to be more
permissive and let go through all of the given table names and clean the
tableKeys cache on each occurrence.