Unnecessary update is treated as an error
There are cases during a write query where part of the data
in a transation does not need to get updated because the data
are unchanged.

Since there are no rows effected the result will be 0.

As per `wpdb::query` documentation, `false` is returned in case of errors,
where `boolean` indicate the number of row effected.

Therefore `!$result` where `$result === 0` is treated like an error and shouldn't
because isn't correct and will make the transaction fails without reason.