Modernized inpsyde/dbal with new coding standards and PHP 8.0 min.
This PR is based on the dev/upgrade-php-wp-modernize branch and merges the latest changes of dev/main. Based on the previous work the branch continues its journey and introduces the following changes:

## removed PHP 7.4 min
This change allows us to use "mixed" typehint for arguments, but also removes the need of some phpcs:disabled lines.

## upgrade to syde/phpcs
The old coding standards inpyde/php-coding-standards have been replaced with syde/phpcs and the whole code base was reformatted and adopted to the new ruleset.

## replace pslam with phpstan
This was the biggest changed introduced in this commit. PSALM was completely removed and PHPStan was introduced with level 6 for now. In future we will increase the PHPStan level to 8.